shadows/requirements.txt
pyglm==1.1.2
noise==1.2.2
PyOpenGL==3.1.0